v1.1.0.0 - User-friendly settings form, retry countdown dialog, PASS/FAIL label, Cancel Test button moved to equipment panel

+++ DuplexerFinalTest/Models/FinalTestSpecModel.cs
@@ -5,7 +5,7 @@ namespace DuplexerFinalTest.Models
         public int TestID { get; set; }
         public string TestCaption { get; set; }
         public string TestUnit { get; set; }
-        public float LimitMin { get; set; }
-        public float LimitMax { get; set; }
+        public double LimitMin { get; set; }
+        public double LimitMax { get; set; }
     }
 }
